requirements.txt
flask
pandas
scikit-learn==1.5.1
flask-cors
pypdf
pytesseract
pdf2image
Pillow
requests
